conda-requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
arviz=0.12.1=pyhd8ed1ab_0
async-generator=1.10=pypi_0
attrs=21.4.0=pypi_0
bcrypt=3.2.2=pypi_0
brotli=1.0.9=h8ffe710_7
brotli-bin=1.0.9=h8ffe710_7
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.5.18.1=h5b45459_0
certifi=2022.5.18.1=py38haa244fe_0
cffi=1.15.0=pypi_0
cftime=1.6.0=py38hbdcd294_1
charset-normalizer=2.0.12=pypi_0
click=8.1.3=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
convertdate=2.4.0=pyhd8ed1ab_0
cryptography=37.0.2=pypi_0
curl=7.83.1=h789b8ee_0
cycler=0.11.0=pyhd8ed1ab_0
cython=0.29.30=py38h885f38d_0
dnspython=2.2.1=pypi_0
email-validator=1.2.1=pypi_0
ephem=4.1.2=py38h2bbff1b_0
et-xmlfile=1.1.0=pypi_0
fbprophet=0.7.1=py38h7ae7562_0
flask=2.1.2=pypi_0
flask-bcrypt=1.0.1=pypi_0
flask-login=0.6.1=pypi_0
flask-sqlalchemy=2.5.1=pypi_0
flask-wtf=1.0.1=pypi_0
fonttools=4.33.3=py38h294d835_0
freetype=2.10.4=h546665d_1
greenlet=1.1.2=pypi_0
h11=0.13.0=pypi_0
hdf4=4.2.15=h0e5069d_3
hdf5=1.12.1=nompi_h2a0e4a3_104
hijri-converter=2.2.4=pyhd8ed1ab_0
holidays=0.13=pyhd8ed1ab_0
idna=3.3=pypi_0
importlib-metadata=4.11.4=py38haa244fe_0
importlib_metadata=4.11.4=hd8ed1ab_0
intel-openmp=2022.1.0=h57928b3_3787
itsdangerous=2.1.2=pypi_0
jinja2=3.1.2=pypi_0
jpeg=9e=h8ffe710_1
kiwisolver=1.4.2=py38hbd9d945_1
korean_lunar_calendar=0.2.1=pyh9f0ad1d_0
krb5=1.19.3=h1176d77_0
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libblas=3.9.0=15_win64_mkl
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libcblas=3.9.0=15_win64_mkl
libcurl=7.83.1=h789b8ee_0
libdeflate=1.10=h8ffe710_0
liblapack=3.9.0=15_win64_mkl
libnetcdf=4.8.1=nompi_h1cc8e9d_102
libpng=1.6.37=h1d00b33_2
libpython=2.2=py38haa244fe_1
libssh2=1.10.0=h680486a_2
libtiff=4.3.0=hc4061b1_4
libwebp=1.2.2=h57928b3_0
libwebp-base=1.2.2=h8ffe710_1
libxcb=1.13=hcd874cb_1004
libzip=1.8.0=hfed4ece_1
libzlib=1.2.12=h8ffe710_0
lunarcalendar=0.0.9=py_0
lz4-c=1.9.3=h8ffe710_1
m2w64-binutils=2.25.1=5
m2w64-bzip2=1.0.6=6
m2w64-crt-git=5.0.0.4636.2595836=2
m2w64-gcc=5.3.0=6
m2w64-gcc-ada=5.3.0=6
m2w64-gcc-fortran=5.3.0=6
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gcc-objc=5.3.0=6
m2w64-gmp=6.1.0=2
m2w64-headers-git=5.0.0.4636.c0ad18a=2
m2w64-isl=0.16.1=2
m2w64-libiconv=1.14=6
m2w64-libmangle-git=5.0.0.4509.2e5a9a2=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
m2w64-make=4.1.2351.a80a8b8=2
m2w64-mpc=1.0.3=3
m2w64-mpfr=3.1.4=4
m2w64-pkg-config=0.29.1=2
m2w64-toolchain=5.3.0=7
m2w64-toolchain_win-64=2.4.0=0
m2w64-tools-git=5.0.0.4592.90b8472=2
m2w64-windows-default-manifest=6.4=3
m2w64-winpthreads-git=5.0.0.4634.697f757=2
m2w64-zlib=1.2.8=10
markupsafe=2.1.1=pypi_0
matplotlib-base=3.5.2=py38he529843_0
mkl=2022.1.0=h6a75c08_874
mpmath=1.2.1=pypi_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
netcdf4=1.5.8=nompi_py38h0500770_101
numpy=1.22.4=py38h1d2777f_0
openjpeg=2.4.0=hb211442_1
openpyxl=3.0.10=pypi_0
openssl=1.1.1o=h8ffe710_0
outcome=1.1.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=py38hcc40339_2
pillow=9.1.1=py38hd8e0db4_0
pip=21.2.2=py38haa95532_0
plotly=5.8.1=pypi_0
pthread-stubs=0.4=hcd874cb_1001
pycparser=2.21=pypi_0
pymeeus=0.5.10=pyhd8ed1ab_0
pyopenssl=22.0.0=pypi_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=pypi_0
pystan=2.19.1.1=py38h02adce6_3
python=3.8.13=h6244533_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=0.20.0=pypi_0
python_abi=3.8=2_cp38
pytz=2022.1=pyhd8ed1ab_0
requests=2.28.0=pypi_0
scipy=1.8.1=py38h9bf8e03_0
selenium=4.2.0=pypi_0
setuptools=61.2.0=py38haa95532_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.2.0=pypi_0
sortedcontainers=2.4.0=pypi_0
sqlalchemy=1.4.37=pypi_0
sqlite=3.38.3=h2bbff1b_0
sympy=1.10.1=pypi_0
tbb=2021.5.0=h2d74725_1
tenacity=8.0.1=pypi_0
tk=8.6.12=h8ffe710_0
tqdm=4.64.0=pyhd8ed1ab_0
trio=0.21.0=pypi_0
trio-websocket=0.9.2=pypi_0
typing_extensions=3.10.0.2=pyha770c72_0
unicodedata2=14.0.0=py38h294d835_1
urllib3=1.26.9=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
webdriver-manager=3.7.0=pypi_0
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py38haa95532_2
wsproto=1.1.0=pypi_0
wtforms=3.0.1=pypi_0
xarray=2022.3.0=pyhd8ed1ab_0
xarray-einstats=0.2.2=pyhd8ed1ab_0
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.5=h62dcd97_1
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.12=h8ffe710_0
zstd=1.5.2=h6255e5f_1
